Return associated expenses in GET /api/trips
Flavio's chosen to run one query per trip with this approach; it's the
old N+1 performance behaviour.

Presumably he's done it this way because we haven't yet covered
relations, and the schema hasn't defined the relationship between Trip
and Expense.

It's an interesting example of how to call await inside a `map()`
though.