Break endpoint docs up into multiple components
This means we can avoid using `&&` to check whether the `parameters` and
`response` properties are set for each endpoint, and can use an `if`
statement instead.

Why can't we put `if (somevar) { ... }` directly inside our JSX? Because
everything that we insert into it needs to "evaluate" to something. It
needs to have a value.

An `if` statement doesn't implicitly return a value, it just sends you
off down a different code path.

On the other hand `somevar && an_expression` will "evaluate to" (i.e. be
equal to) `an_expression` if `somevar` evaluates to true. If `somevar`
isn't "truthy" then it will evaluate to whatever `somevar` is set to,
which in the code that I've just removed, would have been undefined.

JSX will quite happily skip over `undefined` when it encounters it,
rendering no markup at all.